YLW-3314 refactor: Tidies up nullable types.  updateInstrument changed to InstrumentPut
