+++ MusicSchool/Models/Category.cs
@@ -7,5 +7,5 @@ public class Category
     public required int Id { get; set; }
     public required string CategoryName { get; set; }
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
-    public IEnumerable<Instrument>? Instruments { get; }
+    public IEnumerable<Instrument> Instruments { get; } = null!;
 }
