+++ MusicSchool/Responses/InstrumentResponse.cs
@@ -15,9 +15,9 @@ public class InstrumentResponse
     [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
     public IEnumerable<StudentResponse>? Students { get; set; }
 
-    public InstrumentResponse(int id, string name, string categoryName, IEnumerable<Student>? students) : this(id, name, categoryName)
+    public InstrumentResponse(int id, string name, string categoryName, IEnumerable<Student> students) : this(id, name, categoryName)
     {
-        Students = students?.Select(x => new StudentResponse(x.Id, $"{x.FirstName} {x.LastName}", x.DateOfBirth));
+        Students = students.Select(x => new StudentResponse(x.Id, $"{x.FirstName} {x.LastName}", x.DateOfBirth));
     }
 
     public InstrumentResponse(int id, string name, string categoryName) : this(id, name)
